Add support for root-level 404.html, etc. pages
Adds some complexity, since we can no longer assume where a page's .html
and its history.html are in relation to each other.  Nor can we assume
that the page's .html is index.html.

Making [0-9]{3} pages a separate branch entirely would've meant doubling
up a bunch of templates though, so that was a no-go.